Move storage providers to lib
