+++ src/lib/formio.ts
@@ -1,7 +1,7 @@
 import { Formio } from 'formiojs';
 import { Form, FormBuilder } from '@formio/vue';
 
-import BeabeeStorage from '../components/form-builder/providers/storage/beabee';
+import BeabeeStorage from './formio/providers/storage/beabee';
 
 // Formio type is broken
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
